Update type of total_time for enhanced precision in Timer.h
The type of total_time has been changed to 'auto' with the addition of a long double specifier for better precision. This is also done to improve the code readability. Additionally, introduced a 'NOLINTNEXTLINE' comment to avoid static analysis tools from flagging the variable total_time, which is potentially unused. The logic of the code remains unchanged.

Signed-off-by: GIUSEPPE BIANCONI <gbianconi3@gmail.com>