Add NOLINT functionStatic to function declarations in Tokenizer.h
Added NOLINT functionStatic annotation  to isPlusORMinus and isOperator functions in Tokenizer.h to suppress the linter warning about static functions. Previously, the linter raised a warning as these inline methods could potentially become static. The new annotation ensures the linter ignores the potential issue, as the warning is not relevant in this context.

Signed-off-by: GIUSEPPE BIANCONI <gbianconi3@gmail.com>