Add variable length operator support in Tokenizer
This commit adds the ability to handle variable length operators like "+=", "-=", "*=", "/=", ">=", "<=", and "!=" in `Tokenizer.h` and `Tokenizer.cpp`. Also, new token types are introduced in `Token.h` to support these changes. This improvement is important as it increases the versatility of the tokenizer and makes it compatible with a wider range of languages and grammars. at the moment the tokenization doesn't work

Signed-off-by: GIUSEPPE BIANCONI <gbianconi3@gmail.com>