Add boolean operators and refactor log format
Added an array for testing a range of boolean operator codes. Redesigned log format to improve readability: if the code string length is under 93 characters, the length is now logged on the same line as the code. For longer strings, the length is logged separately. These changes aim to improve code testing and debugging efficiency.

Signed-off-by: GIUSEPPE BIANCONI <gbianconi3@gmail.com>