This commit includes a variety of improvements and refactoring changes within the Tokenizer class. Notably, the code received modifications to boost clarity, simplify, and standardize processes, for example, increasing the current position and column of the tokenizer was refactored into its own separate function to reduce code repetition.
Additionally, the documentation was greatly improved for public methods and class descriptions, adhering to the Doxygen style. This will aid both developers working on the code and users of it in understanding the responsibilities and usage of the Tokenizer class.

Minor renaming of parameters and method variables was implemented to enhance code readability and understanding, and comments within the code were also expanded and improved, giving in-depth explanations of complex sections of the code.

Signed-off-by: GIUSEPPE BIANCONI <gbianconi3@gmail.com>