Update comment extraction logic in Tokenizer
Modified the comment extraction method in Tokenizer.cpp to also recognize multi-line comments. In the original version, it was identifying comments only by the presence of "//" or "/*". The function was updated to identify comments by either single line or multi line comment at the beginning, allowing it to better represent the variety of comments that can be found in a codebase.

Signed-off-by: GIUSEPPE BIANCONI <gbianconi3@gmail.com>